+++ CHANGELOG.md
@@ -7,6 +7,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased - xxxx-xx-xx
+
+### Fixed
+
+- WCAG issues Date(Range)Picker: correct aria and role attributes & trap focus inside modal `Calendar` component
+
 ## 7.6.0 - 2024-12-10
 
 ### Added